specify the proper type of versionCode

+++ app/src/main/java/org/grapheneos/apps/client/item/PackageVariant.kt
@@ -5,7 +5,7 @@ data class PackageVariant(
     val pkgName: String,
     val type: String,
     val packagesInfo: Map<String, String>, //package and hash
-    val versionCode: Int,
+    val versionCode: Long,
     val dependencies: List<String> = listOf(),
     val originalPkgName: String? = null
 )
